Feature/376552 generic error screen (#14)
* Feature/task/419381 create generic frontend controller (#9)

* created generic frontend

* added Unit tests for controller, service, RestServices

* Refactoring and bugfix

* Feature/task/422021 create generic error page (#13)

* create generic error page initial checkin

* Controller names were changed

* added Unit tests for Failure page

---------

+++ src/EPR.Payment.Portal.Common/Models/CompletePaymentViewModel.cs
@@ -10,6 +10,6 @@ namespace EPR.Payment.Portal.Common.Models
         public Guid? UserId { get; set; }
         public Guid? OrganisationId { get; set; }
         public string? Regulator { get; set; }
-        public decimal? Amount { get; set; }
+        public int? Amount { get; set; }
     }
 }
